Separate scratchpad
